pyproject.toml
[project]
name = "optimizer-340b"
version = "1.0.0"
description = "Site-of-Care Optimization Engine for 340B drug program revenue maximization"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["340b", "healthcare", "pharmacy", "optimization", "revenue"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "streamlit>=1.30.0",
    "polars>=0.20.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "plotly>=5.18.0",
    "python-dotenv>=1.0.0",
    "orjson>=3.9.0",
    "pydantic>=2.5.0",
    "thefuzz>=0.22.0",
    "python-Levenshtein>=0.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
    "mypy>=1.10",
    "pre-commit>=3.7",
    "pandas-stubs>=2.0.0",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["optimizer_340b"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=src --cov-report=term-missing"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/optimizer_340b"]
